+++ src/main/java/lotto/domain/Lotto.java
@@ -5,8 +5,6 @@ import java.util.Comparator;
 import java.util.HashSet;
 import java.util.List;
 
-import lotto.domain.vo.LottoNumber;
-
 public class Lotto {
 
     private static final int LOTTO_SIZE = 6;
