+++ src/main/java/lotto/domain/LottoGenerator.java
@@ -2,13 +2,11 @@ package lotto.domain;
 
 import java.util.ArrayList;
 import java.util.Collections;
-import java.util.Comparator;
 import java.util.List;
 import java.util.stream.Collectors;
 import java.util.stream.IntStream;
 
 import lotto.constants.NumberLimit;
-import lotto.domain.vo.LottoNumber;
 
 public class LottoGenerator {
