+++ src/main/java/lotto/domain/LottoMoney.java
@@ -1,7 +1,5 @@
-package lotto.domain.vo;
+package lotto.domain;
 
-import java.math.BigDecimal;
-import java.math.RoundingMode;
 import java.text.MessageFormat;
 import java.util.Objects;
