+++ src/main/java/lotto/domain/Rank.java
@@ -1,8 +1,5 @@
 package lotto.domain;
 
-import lotto.domain.vo.LottoMoney;
-import lotto.domain.vo.Reward;
-
 import java.util.function.BiPredicate;
 import java.util.*;
