+++ src/main/java/lotto/domain/Store.java
@@ -1,6 +1,5 @@
 package lotto.domain;
 
-import lotto.domain.vo.LottoMoney;
 import lotto.view.InputView;
 
 import java.util.ArrayList;
