+++ src/main/java/lotto/view/OutputView.java
@@ -7,7 +7,7 @@ import java.util.List;
 import java.util.stream.Collectors;
 
 import lotto.domain.Lotto;
-import lotto.domain.vo.LottoNumber;
+import lotto.domain.LottoNumber;
 import lotto.domain.Rank;
 import lotto.controller.Rate;
