+++ src/test/java/lotto/domain/LottoNumberTest.java
@@ -1,4 +1,4 @@
-package lotto.domain.vo;
+package lotto.domain;
 
 import static org.assertj.core.api.Assertions.*;
